backend_oauth2_local_tokeninfo/requirements.txt
pyjwt
flask-cors
connexion